refactor(components): refactor rate
