+++ packages/components/rate/index.ts
@@ -1,13 +1,8 @@
-import Rate from './src/index.vue'
+import { withInstall } from '@element-plus/utils/with-install'
 
-import type { App } from 'vue'
-import type { SFCWithInstall } from '@element-plus/utils/types'
+import Rate from './src/rate.vue'
 
-Rate.install = (app: App): void => {
-  app.component(Rate.name, Rate)
-}
+export const ElRate = withInstall(Rate)
+export default ElRate
 
-const _Rate = Rate as SFCWithInstall<typeof Rate>
-
-export default _Rate
-export const ElRate = _Rate
+export * from './src/rate'
